重构代码，新增一个动态运行js脚本的方法。
